Move index command to own file
